requirements.txt
# Dev environment
pip
autopep8

# App
google-generativeai
langchain_core
langchain-google-genai # This package contains the LangChain integrations for the Gemini API.
Pillow # PIL image-loading library
